scripts_requirements_py3.txt
avro-python3==1.8.2
kafka-python==1.3.3
mysql-connector==2.2.3
py4j==0.10.4
pypandoc==1.4
pyspark==2.2.0
python-dateutil==2.6.1
six==1.10.0
